+++ HISTORY.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 * Refactor API functions to use specific types
+* Refactor amount buffer types for native and non-native amounts
 
 ## 0.4.0 (2024-9-20)